# src/api/rest-api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorModel": {
        "properties": {
          "error": {
            "description": "Error type",
            "example": "Bad Request",
            "type": "string"
          },
          "message": {
            "description": "Detailed error message",
            "example": "Validation failed",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code",
            "example": 400,
            "type": "number"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "ProjectModel": {
        "properties": {
          "createdAt": {
            "description": "Timestamp when the project was created",
            "example": "2024-01-15T10:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Unique project identifier",
            "example": "V1StGXR8_Z5jdHi6B-myT",
            "type": "string"
          },
          "name": {
            "description": "URL-friendly project name (lowercase letters, numbers, hyphens)",
            "example": "my-blog",
            "type": "string"
          },
          "organizationId": {
            "description": "Organization that owns this project",
            "example": "acme-corp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organizationId",
          "createdAt",
          "name"
        ],
        "type": "object"
      },
...
      "SuccessModelDto": {
        "properties": {
          "success": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
      "UpdateProjectDto": {
        "properties": {
          "isPublic": {
            "type": "boolean"
          }
        },
        "required": [
          "isPublic"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Revisium API",
    "version": "2.6.0-alpha.3"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organization/{organizationId}/projects/{projectName}": {
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "project",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get project by name",
        "tags": [
          "Project"
        ]
      },
      "put": {
        "operationId": "updateProject",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessModelDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update project settings",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/organization/{organizationId}/projects/{projectName}/branches": {
      "get": {
        "operationId": "branches",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "first",
            "required": true,
            "schema": {
              "default": 100,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchesConnection"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all branches in the project",
        "tags": [
          "Project"
        ]
      }
    },
...
    "/api/organization/{organizationId}/projects/{projectName}/branches/{branchName}/revisions": {
      "get": {
        "operationId": "revisions",
        "parameters": [
          {
            "description": "Organization identifier (typically the owner username)",
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "URL-friendly project name (lowercase, numbers, hyphens allowed)",
            "in": "path",
            "name": "projectName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Branch name",
            "in": "path",
            "name": "branchName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "first",
            "required": true,
            "schema": {
              "default": 100,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inclusive",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Sort order: asc (default) or desc",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionsConnection"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Branch not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all revisions in the branch",
        "tags": [
          "Branch"
        ]
      }
    },
...
    "/api/revision/{revisionId}": {
      "get": {
        "operationId": "revision",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get revision by ID",
        "tags": [
          "Revision"
        ]
      }
    },
    "/api/revision/{revisionId}/apply-migrations": {
      "post": {
        "operationId": "applyMigrations",
        "parameters": [
          {
            "description": "Draft revision identifier (modifications are only allowed in draft revisions)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InitMigrationDto"
                    },
                    {
                      "$ref": "#/components/schemas/UpdateMigrationDto"
                    },
                    {
                      "$ref": "#/components/schemas/RenameMigrationDto"
                    },
                    {
                      "$ref": "#/components/schemas/RemoveMigrationDto"
                    }
                  ]
                },
                "type": "array"
              }
            }
          },
          "description": "Array of migration operations to apply",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplyMigrationsResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Apply schema migrations to this revision",
        "tags": [
          "Revision"
        ]
      }
    },
...
    "/api/revision/{revisionId}/child-revision": {
      "get": {
        "operationId": "childRevision",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get child revision",
        "tags": [
          "Revision"
        ]
      }
    },
...
    "/api/revision/{revisionId}/parent-revision": {
      "get": {
        "operationId": "parentRevision",
        "parameters": [
          {
            "description": "Revision identifier (any revision: draft, head, or historical)",
            "in": "path",
            "name": "revisionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionModel"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid request data or validation failed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Authentication required. Provide a valid JWT token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Insufficient permissions for this operation"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Revision not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get parent revision",
        "tags": [
          "Revision"
        ]
      }
    },
...
  },
  "servers": [],
...
}